Moved SearchPageBase into components/BF, refactored unit search page into a more customizable picker component
