+++ src/components/BF/SearchArea.vue
