+++ src/components/BF/SearchPageBase.vue
